Fix a bug that create new character will crash server. add a field bag in table chracter_inventory to save bag index of item.
git-svn-id: http://mangos.svn.sourceforge.net/svnroot/mangos@1159 26674162-90ff-0310-a85b-b64107c8dee9

+++ trunk/src/game/Bag.cpp
@@ -78,21 +78,19 @@ void Bag::Create(uint32 guidlow, uint32 itemid, Player* owner)
 
 void Bag::SaveToDB()
 {
-    Item::SaveToDB();
-    sDatabase.PExecute("DELETE FROM `character_inventory` WHERE `guid` = '%u'",GetGUIDLow());
-    for (int i = 0; i < 20; i++)
+    sDatabase.PExecute("DELETE FROM `character_inventory` WHERE `guid` = '%u' AND `bag` = '%u';", m_owner->GetGUIDLow(), GetSlot());
+    for (uint8 i = 0; i < 20; i++)
     {
         if (m_bagslot[i])
         {
-            sDatabase.PExecute("INSERT INTO `character_inventory` (`guid`,`slot`,`item`,`item_template`) VALUES ('%u', '%u', '%u', '%u');", GetGUIDLow(), i, m_bagslot[i]->GetGUIDLow(), m_bagslot[i]->GetEntry());
+            sDatabase.PExecute("INSERT INTO `character_inventory`  (`guid`,`bag`,`slot`,`item`,`item_template`) VALUES ('%u', '%u', '%u', '%u', '%u');", GetGUIDLow(), GetSlot(), i, m_bagslot[i]->GetGUIDLow(), m_bagslot[i]->GetEntry());
             m_bagslot[i]->SaveToDB();
         }
     }
 }
 
-void Bag::LoadFromDB(uint32 guid, uint32 auctioncheck)
+void Bag::LoadFromDB()
 {
-    Item::LoadFromDB(guid, auctioncheck);
     for (uint8 i = 0; i < 20; i+=2)
     {
         SetUInt64Value(CONTAINER_FIELD_SLOT_1 + (i*2), 0);
@@ -103,23 +101,23 @@ void Bag::LoadFromDB(uint32 guid, uint32 auctioncheck)
         }
     }
 
-    QueryResult *result = sDatabase.PQuery("SELECT * FROM `character_inventory` WHERE `guid` = '%u';", GetGUIDLow());
+    QueryResult *result = sDatabase.PQuery("SELECT * FROM `character_inventory` WHERE `guid` = '%u' AND `bag` = '%u';", m_owner->GetGUIDLow(), GetSlot());
 
     if (result)
     {
         do
         {
             Field *fields = result->Fetch();
-            uint8  slot      = fields[1].GetUInt8();
-            uint32 item_guid = fields[2].GetUInt32();
-            uint32 item_id   = fields[3].GetUInt32();
+            uint8  slot      = fields[2].GetUInt8();
+            uint32 item_guid = fields[3].GetUInt32();
+            uint32 item_id   = fields[4].GetUInt32();
 
             ItemPrototype* proto = objmgr.GetItemPrototype(item_id);
 
             Item *item = NewItemOrBag(proto);
             item->SetOwner(m_owner);
             item->SetSlot(slot);
-            item->LoadFromDB(item_guid, auctioncheck);
+            item->LoadFromDB(item_guid, 1);
             AddItemToBag(slot, item);
         } while (result->NextRow());
 
@@ -136,7 +134,6 @@ void Bag::DeleteFromDB()
             m_bagslot[i]->DeleteFromDB();
         }
     }
-    Item::DeleteFromDB();
 }
 
 uint8 Bag::FindFreeBagSlot()
